add all packages to 2 hw
