--- src/hw2/Comics.java
@@ -1,15 +1,0 @@
-package hw2;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@EqualsAndHashCode(callSuper = true)
-@Data
-@AllArgsConstructor
-
-public class Comics extends Book {
-    private boolean isAdult;
-
-
-}
