--- src/hw2/Workstation.java
@@ -1,12 +1,0 @@
-package hw2;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@EqualsAndHashCode(callSuper = true)
-@Data
-@AllArgsConstructor
-public class Workstation extends Laptop {
-    private String software;
-}
