+++ src/hw2/part2/Laptop.java
@@ -1,0 +1,17 @@
+package hw2.part2;
+
+import hw2.part2.PC;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@EqualsAndHashCode(callSuper = true)
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Laptop extends PC {
+    private String accessories;
+    private int size;
+
+}
