+++ src/hw2/part2/Workstation.java
@@ -1,14 +1,13 @@
-package hw2;
+package hw2.part2;
 
+import hw2.part2.Laptop;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
 
 @EqualsAndHashCode(callSuper = true)
 @Data
 @AllArgsConstructor
-@NoArgsConstructor
-public class Ultrabook extends Laptop {
-    private int butteryLife;
+public class Workstation extends Laptop {
+    private String software;
 }
