+++ src/hw2/part3/Book.java
@@ -1,0 +1,17 @@
+package hw2.part3;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@EqualsAndHashCode(callSuper = true)
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Book extends Papyrus {
+    private int pages;
+    private String genres;
+    private int price;
+
+}
