+++ src/hw2/part4/Employe.java
@@ -1,4 +1,4 @@
-package hw2;
+package hw2.part4;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,6 +13,6 @@ public class Employe {
     private String email;
     private int age;
     private GENDER gender;
-    private Skills [] skills;
+    private Skills[] skills;
     private Car car;
 }
