+++ src/hw2/part4/GENDER.java
@@ -1,4 +1,4 @@
-package hw2;
+package hw2.part4;
 
 public enum GENDER {
     MALE,
