+++ src/hw2/part4/Skills.java
@@ -1,4 +1,4 @@
-package hw2;
+package hw2.part4;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
